requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
certifi==2018.1.18        # via requests
cffi==1.11.5              # via cryptography
chardet==3.0.4            # via requests
contextlib2==0.5.5        # via raven
cryptography==2.1.4       # via pyopenssl, requests
directory-healthcheck==0.4.2
dj-database-url==0.4.1
django-admin-ip-restrictor==0.1.0
django-environ==0.4.5
django-filter==1.0.4
django-health-check==3.0.0  # via directory-healthcheck
django-ipware==2.0.1      # via django-admin-ip-restrictor
django-pglocks==1.0.2
django==1.11.14
django_extensions==1.6.7
djangorestframework==3.4.7
idna==2.6                 # via cryptography, requests
mohawk==0.3.4             # via sigauth
psycopg2==2.6.1
pycparser==2.18           # via cffi
pyopenssl==17.5.0         # via requests
pytz==2017.2              # via directory-healthcheck, django
raven==5.19.0
requests[security]==2.18.4
sigauth==3.0.0
six==1.11.0               # via cryptography, django-extensions, mohawk, pyopenssl
urllib3==1.22             # via requests
waitress==1.1.0
whitenoise==3.2.2
